feat: comprehensive codebase cleanup and refactoring
- Remove backup files and unnecessary documentation
- Create shared utilities: formatCurrency function with proper negative sign placement
- Create constants: EXPENSE_CATEGORIES with TypeScript typing
- Extract ExpensesList component (230+ lines) into separate modular file
- Update all components to use shared utilities and constants
- Refactor ExpenseForm.tsx to remove embedded ExpensesList component
- Update BalanceSummary.tsx to use shared formatCurrency utility
- Improve code organization and maintainability
- Ensure TypeScript compilation and build success

This refactoring separates concerns, reduces code duplication, and creates
a more maintainable codebase structure following React best practices.
