+++ frontend/src/constants/expenseCategories.ts
@@ -1,0 +1,11 @@
+export const EXPENSE_CATEGORIES = [
+  { value: 'food', label: 'Food & Dining' },
+  { value: 'transportation', label: 'Transportation' },
+  { value: 'accommodation', label: 'Accommodation' },
+  { value: 'entertainment', label: 'Entertainment' },
+  { value: 'shopping', label: 'Shopping' },
+  { value: 'utilities', label: 'Utilities' },
+  { value: 'other', label: 'Other' }
+] as const;
+
+export type ExpenseCategoryValue = typeof EXPENSE_CATEGORIES[number]['value'];
